% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{AnnoGene}
\alias{AnnoGene}
\title{AnnoGene}
\usage{
AnnoGene(
  gr,
  geneCol = "gene_name",
  genes = NULL,
  strandCol = NULL,
  color = "gray30",
  colorCol = NULL,
  shape = "rect"
)
}
\description{
Draws grouped gene annotations as exon boxes connected by multi-arrow lines
}
