% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{AnnotateSvWithSegmentCn}
\alias{AnnotateSvWithSegmentCn}
\title{AnnotateSvWithSegmentCn}
\usage{
AnnotateSvWithSegmentCn(sv_gr, cn_gr, slop = 5000, cn_col = "copy_number")
}
\description{
Annotate an SV vcf with total copy number at breakend positions.
}
