% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{clipToXscale}
\alias{clipToXscale}
\title{clipToXscale}
\usage{
clipToXscale(x0, x1, xscale)
}
\arguments{
\item{x0, x1, xscale}{Coordinates for clipping}
}
\description{
Clipping for ranges intersected by x axis limits
}
