% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{convertDataToGrid}
\alias{convertDataToGrid}
\title{convertDataToGrid}
\usage{
convertDataToGrid(x, y, xscale, yscale)
}
\arguments{
\item{x, y, xscale, yscale}{Coordinates for clipping}
}
\description{
Converts data coordinates to grid coordinates
}
