% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{CreateSequenceWindows}
\alias{CreateSequenceWindows}
\title{CreateSequenceWindows}
\usage{
CreateSequenceWindows(regions, padding = 0, genome = "hg38", add_chr = FALSE)
}
\description{
Create a GRanges object with specified ranges using e.g., chr1:100-200
}
