% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{defaultGenomeWindows}
\alias{defaultGenomeWindows}
\title{defaultGenomeWindows}
\usage{
defaultGenomeWindows(add_chr = T)
}
\description{
Generates a GRanges object of all standard hg38 chromosomes to serve as the default plotting windows.
}
\author{
Andrew Lynch
}
