% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{.recycled_to_N}
\alias{.recycled_to_N}
\title{Recycle/trim a vector to length \code{N}}
\usage{
.recycled_to_N(x, N, default)
}
\arguments{
\item{x}{Vector.}

\item{N}{Target length.}

\item{default}{Default value if \code{x} is \code{NULL} or length 0.}
}
\value{
Vector of length \code{N}.
}
\description{
Recycle/trim a vector to length \code{N}
}
\keyword{internal}
