% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawAnnoFeature}
\alias{drawAnnoFeature}
\title{drawAnnoFeature}
\usage{
drawAnnoFeature(
  gr,
  labels = NULL,
  orientation = NULL,
  shape = "rect",
  x0 = NULL,
  x1 = NULL,
  xlim = NULL,
  ybase = 0,
  tier_height = 1,
  y0s = NULL,
  y1s = NULL,
  col = "gray30",
  border = NA,
  label_cex = 0.6,
  label_offset = 4,
  returnTiers = FALSE
)
}
\description{
Draw annotation features with different shape styles
}
