% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawAnnoGene}
\alias{drawAnnoGene}
\title{drawAnnoGene}
\usage{
drawAnnoGene(
  ag,
  layout,
  track_idx,
  track_y_min = 0,
  track_y_max = 1,
  totalTracks,
  arrow_spacing = 10,
  arrowhead_len = 1,
  exon_height_ratio = 0.8,
  label_cex = 0.7,
  label_offset = 5
)
}
\description{
Draws grouped gene annotations as exon boxes connected by multi-arrow lines

Draws grouped gene annotations as exon boxes connected by multi-arrow lines
}
