% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawArc}
\alias{drawArc}
\title{drawArc}
\usage{
drawArc(x0, y0, x1, y1, orientation = "*", col = "black", lwd = 2)
}
\arguments{
\item{x0, }{y0 Start coordinates}

\item{x1, }{y1 End coordinates}

\item{orientation}{Arc orientation: "*", "up", "down", "+", "-", etc.}

\item{col}{Color}

\item{lwd}{Line width}
}
\description{
Draws a cubic Bezier arc between two points.
}
