% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawIdeogram}
\alias{drawIdeogram}
\title{drawIdeogram}
\usage{
drawIdeogram(feat, layout, track_idx)
}
\arguments{
\item{feat}{A SeqIdeogram object.}

\item{layout}{Layout object from setupGlobalLayout().}

\item{track_idx}{Integer index of the current track.}
}
\value{
None. Draws directly to the active plot.
}
\description{
Draws an ideogram track using cytoband rectangles.
}
