% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawSeqArch}
\alias{drawSeqArch}
\title{drawSeqArch}
\usage{
drawSeqArch(
  x0,
  y0,
  x1,
  y1,
  top0,
  top1,
  orientation = "*",
  curve = "length",
  stemWidth = 1,
  arcWidth = 1,
  arcColor = "black",
  stemColor = "black"
)
}
\arguments{
\item{x0, x1, y0, y1, top0, top1}{Canvas coordinates. y0 and y1 are bottoms of arch stems. top0 and top1 are the tops of arch stems}

\item{orientation}{Orientation of the link. In general, "-" arches will point down, "+" will point up.}

\item{curve}{Argument to set curve type or degree. Can be one of  "length", "equal", or number to control height of arc (default is 0.2).}

\item{stemWidth, arcWidth, arcColor, stemColor}{Aesthetic control over arches.}
}
\description{
General function for drawing SeqLink arches
}
