% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawSeqString}
\alias{drawSeqString}
\title{drawSeqString}
\usage{
drawSeqString(
  x0,
  y0,
  x1,
  y1,
  orientation = "*",
  col = "black",
  lwd = 1.5,
  buffer = NULL
)
}
\description{
Draws a 6-point cubic Bézier DNA string curve
}
