% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{flexoki_palette}
\alias{flexoki_palette}
\title{Load colors from Kepano's Flexoki palette}
\usage{
flexoki_palette(n, shade = 400)
}
\arguments{
\item{n}{Number of colors to be produced. If > 9, an equally distributed color ramp will be produced (and probably look bad).}
}
\value{
Vector of hex colors
}
\description{
Loads colors from Kepano's Flexoki palette.
}
