% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{getGencode}
\alias{getGencode}
\title{Load cytoband table for ideogram rendering}
\usage{
getGencode(genome = "hg38", version = "v40")
}
\arguments{
\item{path}{Optional path to a UCSC-style cytoband table (TSV format).}
}
\value{
A data.frame with cytoband annotations.
}
\description{
This function loads a UCSC-style cytoband file for ideogram drawing.
If no path is given, it loads the built-in cytoband_hg38 dataset from the package.
}
