% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{globalTransformX}
\alias{globalTransformX}
\title{Transform Genomic X Coordinates to Absolute Plot X Coordinates}
\usage{
globalTransformX(x, w, layout)
}
\arguments{
\item{x}{A numeric vector of genomic positions.}

\item{w}{An integer vector of window indices corresponding to each genomic position in x.}

\item{layout}{A list returned by setupGlobalLayout() that includes global_window_starts, global_window_ends,
and global_windows.}
}
\value{
A numeric vector of absolute x coordinates.
}
\description{
Converts genomic x coordinates into absolute plot x coordinates using the layout information.
}
