% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+\%}
\alias{\%+\%}
\title{Add a SeqElement to a SeqTrack or SeqPlot}
\usage{
e1 \%+\% e2
}
\arguments{
\item{e1}{A `SeqPlot` or `SeqTrack`.}

\item{e2}{A `SeqElement` or `SeqLink`.}
}
\value{
`e1`, invisibly modified in place.
}
\description{
When applied to a `SeqPlot`, `%+%` appends the element to the last (current)
track.  When applied to a `SeqTrack` directly, it appends the element to
that track.
}
