% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{gridToCanvas}
\alias{gridToCanvas}
\title{gridToCanvas}
\usage{
gridToCanvas(u, v, panel_meta)
}
\arguments{
\item{u, v, panel_meta}{Coordinates for clipping}
}
\description{
Converts grid coordinates to canvas coordinates
}
