% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{makeRandomHg38Grange}
\alias{makeRandomHg38Grange}
\title{makeRandomHg38Grange}
\usage{
makeRandomHg38Grange(n_intervals = 1000, min_width = 50, max_width = 1000)
}
\description{
Generate a random hg38 grange object
}
