% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{positionToArm}
\alias{positionToArm}
\title{Convert chromosome position to chromosome arm}
\usage{
positionToArm(chr, pos, cyto = cytoband_hg38)
}
\arguments{
\item{chr}{Chromosome}

\item{pos}{Position}
}
\value{
Character vector of chromosome arms
}
\description{
This function converts chromosome position to chromosome arm
}
