% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqAnnotation-class}
\alias{SeqAnnotation-class}
\title{SeqAnnotation-Class}
\description{
Generic annotation class for drawing composite genomic annotations (e.g., ideograms, gene models).

Create a SeqAnnotation object.
}
\section{Slots}{

\describe{
\item{\code{annotationType}}{Type of annotation (e.g., "ideogram").}

\item{\code{data}}{Internal data required for rendering the annotation.}
}}

