% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqArc}
\alias{SeqArc}
\title{SeqArc}
\usage{
SeqArc(
  gr1,
  gr2,
  color = "black",
  colorCol = NULL,
  y = NA_real_,
  yCol = NULL,
  y0 = NA_real_,
  y0Col = NULL,
  y1 = NA_real_,
  y1Col = NULL,
  t0 = 0,
  t1 = 0,
  orientation = "*"
)
}
\description{
Constructs a SeqArc object for curved connections.
}
