% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqHighlightClass}
\alias{SeqHighlightClass}
\title{SeqHighlight R6 Class}
\description{
Draw a filled polygon "connector" that highlights the same genomic interval
between two tracks (e.g. an ideogram track and a zoomed view track).
}
\details{
`SeqHighlight` projects each interval in `gr` onto track `t0` and `t1`,
then draws a quadrilateral between the two projected spans.
}
\section{Super classes}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{\link[THEfunc:SeqLink]{THEfunc::SeqLink}} -> \code{SeqHighlight}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqHighlight-new}{\code{SeqHighlightClass$new()}}
\item \href{#method-SeqHighlight-prep}{\code{SeqHighlightClass$prep()}}
\item \href{#method-SeqHighlight-draw}{\code{SeqHighlightClass$draw()}}
\item \href{#method-SeqHighlight-clone}{\code{SeqHighlightClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqHighlight-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqHighlight-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqHighlightClass$new(
  gr,
  t0 = 1,
  t1 = 2,
  side = c("auto", "top", "bottom"),
  fill = "black",
  alpha = 0.15,
  col = NA,
  lwd = 0.5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{A `GRanges` of intervals to highlight.}

\item{\code{t0}}{Track index for the first projection.}

\item{\code{t1}}{Track index for the second projection.}

\item{\code{side}}{Which edge of each track to attach to: "auto", "top", "bottom".
If "auto", uses bottom edge of the upper track and top edge of the lower track.}

\item{\code{fill}}{Fill color.}

\item{\code{alpha}}{Fill alpha in [0,1].}

\item{\code{col}}{Border color (NA for no border).}

\item{\code{lwd}}{Border width.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqHighlight-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqHighlight-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqHighlightClass$prep(
  layout_all_tracks,
  track_windows_list,
  arc_track_idx = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqHighlight-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqHighlight-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqHighlightClass$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqHighlight-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqHighlight-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqHighlightClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
