% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{stainToColor}
\alias{stainToColor}
\title{stainToColor}
\usage{
stainToColor(stains)
}
\arguments{
\item{stains}{A character vector of gieStain values.}
}
\value{
A character vector of color hex codes.
}
\description{
Converts UCSC-style cytoband stains into plot colors.
}
