% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{CreateSequenceWindows}
\alias{CreateSequenceWindows}
\title{CreateSequenceWindows}
\usage{
CreateSequenceWindows(regions, padding = 0, genome = "hg38", add_chr = TRUE)
}
\arguments{
\item{regions}{Vector of region strings such as "1:300-4000" or "chr2:20000-40000" or simply "chr3.}

\item{padding}{Number of base pairs around given positions to add. Automatically clipped to chromosome start positions. Default: 0}

\item{genome}{Genome of origin. Default: hg38}

\item{add_chr}{Whether to maintain "chr" convention on seqnames. Default: TRUE}
}
\description{
Generates a GRanges object of given sequence ranges
}
\author{
Andrew Lynch
}
