% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{CreateSequenceWindows}
\alias{CreateSequenceWindows}
\title{CreateSequenceWindows}
\usage{
CreateSequenceWindows(regions, padding = 0, genome = "hg38", add_chr = FALSE)
}
\description{
Generates a GRanges object of given sequence ranges
}
\author{
Andrew Lynch
}
