% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{drawSeqArch}
\alias{drawSeqArch}
\title{drawSeqArch}
\usage{
drawSeqArch(
  x0,
  y0,
  x1,
  y1,
  top0,
  top1,
  orientation = "*",
  curve = "length",
  col = "black",
  lwd = 2
)
}
\arguments{
\item{x0, }{y0 Start base (bottom of left stem)}

\item{x1, }{y1 End base (bottom of right stem)}

\item{top0, }{top1 Absolute Y positions of the top of the stems (in arc track)}

\item{orientation}{Character: "up", "down", "+", "-", "*", etc.}

\item{col}{Color}

\item{lwd}{Line width}
}
\description{
Draws a bounded arch with vertical lines and a curved top.
}
