% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{getGencode}
\alias{getGencode}
\title{Load GTF file for gene tracks}
\usage{
getGencode(genome = "hg38", version = "v32", path = NULL)
}
\arguments{
\item{path}{Optional path to a Gencode-style gene table (GTF format).}
}
\value{
A GRange object with gene annotations
}
\description{
This function loads a Gencode GTF file for drawing gene tracks
If no path is given, it will attempt to load from the database (connection required)
}
