% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{loadCytobands}
\alias{loadCytobands}
\title{Load cytoband table}
\usage{
loadCytobands(path = NULL)
}
\arguments{
\item{path}{Optional path to a UCSC-style cytoband table (TSV format).}
}
\value{
A data.frame with cytoband annotations.
}
\description{
This function loads a UCSC-style cytoband file.
If no path is given, it loads the built-in cytoband_hg38 dataset from the package.
}
