% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqArch}
\alias{SeqArch}
\title{SeqArch}
\usage{
SeqArch(
  gr1,
  gr2,
  color = "black",
  colorCol = NULL,
  y0 = 0,
  y0Col = NULL,
  y1 = 0,
  y1Col = NULL,
  t0 = 0,
  t1 = 0,
  height = 1,
  orientation = "*",
  curve = "length"
)
}
\description{
Constructs a SeqArch object with vertical lines and an arc above.
}
