% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqArch}
\alias{SeqArch}
\title{SeqArch Class}
\format{
A \code{SeqArch} object with the following fields:
}
\value{
A new \code{SeqArch} object.
}
\description{
TODO: Describe the SeqArch class.
}
\section{Methods}{

\describe{
  \item{\code{prep(layout_all_tracks, track_windows_list, arc_track_idx)}}{TODO: Describe 'prep' method.}
}
}

\section{Super classes}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{\link[THEfunc:SeqLink]{THEfunc::SeqLink}} -> \code{SeqArch}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{arcColor}}{TODO: Describe the 'arcColor' field.}

\item{\code{arcWidth}}{TODO: Describe the 'arcWidth' field.}

\item{\code{curve}}{TODO: Describe the 'curve' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}

\item{\code{left_only}}{TODO: Describe the 'left_only' field.}

\item{\code{plotStubs}}{TODO: Describe the 'plotStubs' field.}

\item{\code{right_only}}{TODO: Describe the 'right_only' field.}

\item{\code{stemColor}}{TODO: Describe the 'stemColor' field.}

\item{\code{stemWidth}}{TODO: Describe the 'stemWidth' field.}

\item{\code{stubAngle}}{TODO: Describe the 'stubAngle' field.}

\item{\code{stubLength}}{TODO: Describe the 'stubLength' field.}

\item{\code{stubs}}{TODO: Describe the 'stubs' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arcColor}}{TODO: Describe the 'arcColor' field.}

\item{\code{arcWidth}}{TODO: Describe the 'arcWidth' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}

\item{\code{plotStubs}}{TODO: Describe the 'plotStubs' field.}

\item{\code{stemColor}}{TODO: Describe the 'stemColor' field.}

\item{\code{stemWidth}}{TODO: Describe the 'stemWidth' field.}

\item{\code{stubAngle}}{TODO: Describe the 'stubAngle' field.}

\item{\code{stubLength}}{TODO: Describe the 'stubLength' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqArch-new}{\code{SeqArch$new()}}
\item \href{#method-SeqArch-prep}{\code{SeqArch$prep()}}
\item \href{#method-SeqArch-draw}{\code{SeqArch$draw()}}
\item \href{#method-SeqArch-clone}{\code{SeqArch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArch-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArch-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArch$new(
  gr1,
  gr2,
  t0 = 0,
  t1 = 0,
  y0 = 0,
  y1 = 0,
  yCol = NULL,
  orientation = "*",
  height = 1,
  curve = "length",
  aesthetics = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArch-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArch-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArch$prep(layout_all_tracks, track_windows_list, arc_track_idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArch-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArch-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArch$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
