% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqArea}
\alias{SeqArea}
\title{SeqArea Class}
\format{
A \code{SeqArea} object with the following fields:
}
\value{
A new \code{SeqArea} object.
}
\description{
TODO: Describe the SeqArea class.
}
\section{Methods}{

\describe{
  \item{\code{new(gr, yCol, groupCol, groupLevels, aesthetics)}}{Create a new \code{SeqArea} object.}
  \item{\code{prep(layout_track, track_windows)}}{TODO: Describe 'prep' method.}
  \item{\code{draw()}}{TODO: Describe 'draw' method.}
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqArea}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{alpha}}{TODO: Describe the 'alpha' field.}

\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{defaultAesthetics}}{TODO: Describe the 'defaultAesthetics' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}

\item{\code{group}}{TODO: Describe the 'group' field.}

\item{\code{groupCol}}{TODO: Describe the 'groupCol' field.}

\item{\code{groupLevels}}{TODO: Describe the 'groupLevels' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{stringsAsFactors}}{TODO: Describe the 'stringsAsFactors' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}

\item{\code{yStackedMax}}{TODO: Describe the 'yStackedMax' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{alpha}}{TODO: Describe the 'alpha' field.}

\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{stringsAsFactors}}{TODO: Describe the 'stringsAsFactors' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqArea-new}{\code{SeqArea$new()}}
\item \href{#method-SeqArea-prep}{\code{SeqArea$prep()}}
\item \href{#method-SeqArea-draw}{\code{SeqArea$draw()}}
\item \href{#method-SeqArea-clone}{\code{SeqArea$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArea$new(
  gr,
  yCol = NULL,
  groupCol = NULL,
  groupLevels = NULL,
  aesthetics = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{TODO: Describe parameter 'gr'.}

\item{\code{yCol}}{TODO: Describe parameter 'yCol'.}

\item{\code{groupCol}}{TODO: Describe parameter 'groupCol'.}

\item{\code{groupLevels}}{TODO: Describe parameter 'groupLevels'.}

\item{\code{aesthetics}}{TODO: Describe parameter 'aesthetics'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArea$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArea$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqArea-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqArea-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqArea$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
