% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqBar}
\alias{SeqBar}
\title{SeqBar}
\usage{
SeqBar(gr, yCol = NULL, y1Col = NULL, colorCol = NULL, color = "black")
}
\arguments{
\item{gr}{GRanges object}

\item{yCol}{Metadata column for bottom y-value}

\item{y1Col}{Metadata column for top y-value}

\item{colorCol}{Metadata column for color (optional)}

\item{color}{Default color}
}
\description{
Constructs a SeqBar object for plotting rectangular blocks.
}
