% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqElement}
\alias{SeqElement}
\title{SeqElement Class}
\format{
A \code{SeqElement} object with the following fields:
}
\value{
A new \code{SeqElement} object.
}
\description{
TODO: Describe the SeqElement class.
}
\section{Methods}{

\describe{
  \item{\code{new(gr, yCol)}}{Create a new \code{SeqElement} object.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coordGrid}}{TODO: Describe the 'coordGrid' field.}

\item{\code{coordOriginal}}{TODO: Describe the 'coordOriginal' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqElement-new}{\code{SeqElement$new()}}
\item \href{#method-SeqElement-clone}{\code{SeqElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqElement-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqElement-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqElement$new(gr, yCol = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{TODO: Describe parameter 'gr'.}

\item{\code{yCol}}{TODO: Describe parameter 'yCol'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
