% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{SeqElement}
\alias{SeqElement}
\title{SeqElement R6 Class}
\description{
Base R6 class representing a generic sequence element in the SeqPlot
framework. Stores the original genomic ranges and provides slots for
transformed coordinates in grid space.
}
\details{
The `SeqElement` class is the root R6 class for all genomic elements
in the plotting system. It is initialized with a `GRanges` object and
retains both the original genomic coordinates (`coordOriginal`) and
transformed grid coordinates (`coordGrid`).
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(1:5, width = 1))
el <- SeqElement$new(gr)
el$gr

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing the genomic coordinates
and any associated metadata.}

\item{\code{coordOriginal}}{A `GRanges` object storing the unmodified input
coordinates.}

\item{\code{coordGrid}}{Placeholder for transformed coordinates after applying
the SeqPlot grid layout.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqElement-new}{\code{SeqElement$new()}}
\item \href{#method-SeqElement-clone}{\code{SeqElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqElement-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqElement-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SeqElement` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqElement$new(gr, yCol = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing genomic positions and optional
metadata columns.}

\item{\code{yCol}}{Optional character string naming a metadata column in `gr`
to be used for y-axis values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SeqElement` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
