% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqGene}
\alias{SeqGene}
\title{SeqGene Class}
\format{
A \code{SeqGene} object with the following fields:
}
\value{
A new \code{SeqGene} object.
}
\description{
TODO: Describe the SeqGene class.
}
\section{Methods}{

\describe{
  \item{\code{prep(layout_track, track_windows)}}{TODO: Describe 'prep' method.}
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqGene}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arrow_len}}{TODO: Describe the 'arrow_len' field.}

\item{\code{arrow_spacing}}{TODO: Describe the 'arrow_spacing' field.}

\item{\code{color}}{TODO: Describe the 'color' field.}

\item{\code{colorCol}}{TODO: Describe the 'colorCol' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{exon_height}}{TODO: Describe the 'exon_height' field.}

\item{\code{geneCol}}{TODO: Describe the 'geneCol' field.}

\item{\code{genesFilter}}{TODO: Describe the 'genesFilter' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}

\item{\code{label_cex}}{TODO: Describe the 'label_cex' field.}

\item{\code{label_offset}}{TODO: Describe the 'label_offset' field.}

\item{\code{label_pad}}{TODO: Describe the 'label_pad' field.}

\item{\code{shape}}{TODO: Describe the 'shape' field.}

\item{\code{strandCol}}{TODO: Describe the 'strandCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{initialize}}{TODO: Describe the 'initialize' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqGene-new}{\code{SeqGene$new()}}
\item \href{#method-SeqGene-prep}{\code{SeqGene$prep()}}
\item \href{#method-SeqGene-draw}{\code{SeqGene$draw()}}
\item \href{#method-SeqGene-clone}{\code{SeqGene$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqGene-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqGene-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqGene$new(
  gr,
  geneCol = "gene_name",
  genesFilter = NULL,
  strandCol = NULL,
  color = "gray30",
  colorCol = NULL,
  shape = "rect"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqGene-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqGene-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqGene$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqGene-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqGene-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqGene$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqGene-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqGene-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqGene$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
