% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqIdeogram}
\alias{SeqIdeogram}
\title{SeqIdeogram Class}
\format{
A \code{SeqIdeogram} object with the following fields:
}
\value{
A new \code{SeqIdeogram} object.
}
\description{
TODO: Describe the SeqIdeogram class.
}
\section{Methods}{

\describe{
  \item{\code{new(cytobands)}}{Create a new \code{SeqIdeogram} object.}
  \item{\code{prep(layout_track, track_windows)}}{TODO: Describe 'prep' method.}
  \item{\code{draw()}}{TODO: Describe 'draw' method.}
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqIdeogram}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{centroPolys}}{TODO: Describe the 'centroPolys' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{cytobands}}{TODO: Describe the 'cytobands' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{stringsAsFactors}}{TODO: Describe the 'stringsAsFactors' field.}

\item{\code{width}}{TODO: Describe the 'width' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{stringsAsFactors}}{TODO: Describe the 'stringsAsFactors' field.}

\item{\code{width}}{TODO: Describe the 'width' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqIdeogram-new}{\code{SeqIdeogram$new()}}
\item \href{#method-SeqIdeogram-prep}{\code{SeqIdeogram$prep()}}
\item \href{#method-SeqIdeogram-draw}{\code{SeqIdeogram$draw()}}
\item \href{#method-SeqIdeogram-clone}{\code{SeqIdeogram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$new(cytobands)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cytobands}}{TODO: Describe parameter 'cytobands'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
