% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{SeqIdeogram}
\alias{SeqIdeogram}
\title{SeqIdeogram R6 Class}
\description{
An R6 class for drawing chromosome ideograms from cytogenetic band
annotations. Each band is drawn as a rectangle, and centromeres
(`acen` regions) are rendered as paired red triangles.
}
\examples{

## ------------------------------------------------
## Method `SeqIdeogram$new`
## ------------------------------------------------

cb <- GenomicRanges::GRanges(
  seqnames = Rle("chr1"),
  ranges   = IRanges::IRanges(c(1, 500001, 1000001), width = 500000),
  gieStain = c("gneg", "acen", "gpos100")
)
ideogram <- SeqIdeogram$new(cb)
}
\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqIdeogram}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cytobands}}{A `GRanges` object of cytogenetic bands. Must
contain a `gieStain` metadata column.}

\item{\code{coordCanvas}}{A list of data frames holding canvas
coordinates for non-centromeric bands (`x0`, `x1`, `y0`, `y1`)
and their fill colors.}

\item{\code{centroPolys}}{A list of polygons representing centromere
regions, each stored as x/y coordinate vectors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqIdeogram-new}{\code{SeqIdeogram$new()}}
\item \href{#method-SeqIdeogram-prep}{\code{SeqIdeogram$prep()}}
\item \href{#method-SeqIdeogram-draw}{\code{SeqIdeogram$draw()}}
\item \href{#method-SeqIdeogram-clone}{\code{SeqIdeogram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SeqIdeogram` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$new(cytobands)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cytobands}}{A `GRanges` object of cytobands with a `gieStain`
metadata column.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SeqIdeogram` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cb <- GenomicRanges::GRanges(
  seqnames = Rle("chr1"),
  ranges   = IRanges::IRanges(c(1, 500001, 1000001), width = 500000),
  gieStain = c("gneg", "acen", "gpos100")
)
ideogram <- SeqIdeogram$new(cb)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-prep}{}}}
\subsection{Method \code{prep()}}{
Prepare ideogram band and centromere coordinates for each panel
window in the SeqPlot layout.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{A list of panel metadata from
`SeqPlot$layoutGrid()`.}

\item{\code{track_windows}}{A `GRanges` object of genomic windows for the
track.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the ideogram for all windows: non-centromeric bands as
rectangles, centromeres as paired red triangles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqIdeogram-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqIdeogram-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqIdeogram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
