% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot-R6.R
\name{SeqLine}
\alias{SeqLine}
\title{SeqLine R6 Class}
\description{
R6 class for drawing line plots from genomic intervals in the SeqPlot
framework. Inherits from [SeqElement].
}
\details{
Each genomic interval is represented as a point at its midpoint by default,
with y-values taken from a metadata column or set to a constant. The points
are connected into a continuous line. Step lines can also be drawn by setting
the `type` aesthetic to `"s"` or `"step"`.
}
\examples{
gr <- GenomicRanges::GRanges(
  "chr1",
  IRanges::IRanges(c(1, 100, 200), width = 50),
  score = c(2, 5, 3)
)
line <- SeqLine$new(gr, yCol = "score")
line$prep(layout_track = some_layout, track_windows = some_windows)
line$draw()

}
\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqLine}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing genomic intervals.}

\item{\code{y}}{Numeric vector of y-values (from `yCol` or constant).}

\item{\code{yCol}}{Optional column name in `gr` used for y-values.}

\item{\code{coordOriginal}}{A `GRanges` object storing unmodified input coordinates.}

\item{\code{coordCanvas}}{List of matrices storing transformed line coordinates
in canvas space for each genomic window.}

\item{\code{aesthetics}}{List of current aesthetics merged with defaults.}

\item{\code{defaultAesthetics}}{Default aesthetics for lines:
\code{type = "n"}, \code{size = 0.1}, \code{color = "#1C1B1A"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqLine-new}{\code{SeqLine$new()}}
\item \href{#method-SeqLine-prep}{\code{SeqLine$prep()}}
\item \href{#method-SeqLine-draw}{\code{SeqLine$draw()}}
\item \href{#method-SeqLine-clone}{\code{SeqLine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SeqLine` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$new(gr, yCol = NULL, aesthetics = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing genomic intervals.}

\item{\code{yCol}}{Optional column name in `gr` for y-values.}

\item{\code{aesthetics}}{Optional list of aesthetic overrides.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SeqLine` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-prep}{}}}
\subsection{Method \code{prep()}}{
Prepare line coordinates in canvas space for each genomic window.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{A list of panel layout metadata for a track.}

\item{\code{track_windows}}{A `GRanges` object defining genomic windows.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-draw}{}}}
\subsection{Method \code{draw()}}{
Draw lines onto the plotting canvas.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
