% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqLine}
\alias{SeqLine}
\title{SeqLine Class}
\format{
A \code{SeqLine} object with the following fields:
}
\value{
A new \code{SeqLine} object.
}
\description{
TODO: Describe the SeqLine class.
}
\section{Methods}{

\describe{
  \item{\code{new(gr, yCol, aesthetics)}}{Create a new \code{SeqLine} object.}
  \item{\code{prep(layout_track, track_windows)}}{TODO: Describe 'prep' method.}
  \item{\code{draw()}}{TODO: Describe 'draw' method.}
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqLine}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{color}}{TODO: Describe the 'color' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{coordOriginal}}{TODO: Describe the 'coordOriginal' field.}

\item{\code{defaultAesthetics}}{TODO: Describe the 'defaultAesthetics' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}

\item{\code{size}}{TODO: Describe the 'size' field.}

\item{\code{type}}{TODO: Describe the 'type' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{color}}{TODO: Describe the 'color' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{size}}{TODO: Describe the 'size' field.}

\item{\code{type}}{TODO: Describe the 'type' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqLine-new}{\code{SeqLine$new()}}
\item \href{#method-SeqLine-prep}{\code{SeqLine$prep()}}
\item \href{#method-SeqLine-draw}{\code{SeqLine$draw()}}
\item \href{#method-SeqLine-clone}{\code{SeqLine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$new(gr, yCol = NULL, aesthetics = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{TODO: Describe parameter 'gr'.}

\item{\code{yCol}}{TODO: Describe parameter 'yCol'.}

\item{\code{aesthetics}}{TODO: Describe parameter 'aesthetics'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
