% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqLink}
\alias{SeqLink}
\title{SeqLink Class}
\format{
A \code{SeqLink} object with the following fields:
}
\value{
A new \code{SeqLink} object.
}
\description{
TODO: Describe the SeqLink class.
}
\section{Methods}{

\describe{
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqLink}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{color}}{TODO: Describe the 'color' field.}

\item{\code{gr1}}{TODO: Describe the 'gr1' field.}

\item{\code{gr2}}{TODO: Describe the 'gr2' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}

\item{\code{orientation}}{TODO: Describe the 'orientation' field.}

\item{\code{t0}}{TODO: Describe the 't0' field.}

\item{\code{t1}}{TODO: Describe the 't1' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{color}}{TODO: Describe the 'color' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqLink-new}{\code{SeqLink$new()}}
\item \href{#method-SeqLink-clone}{\code{SeqLink$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLink-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLink-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLink$new(
  gr1,
  gr2,
  t0 = 0,
  t1 = 0,
  y0 = 0,
  y1 = 0,
  color = "black",
  orientation = "*"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqLink-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqLink-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqLink$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
