% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqMatrix}
\alias{SeqMatrix}
\title{SeqMatrix R6 Class}
\description{
An R6 class for plotting 2D genomic heatmaps (e.g. Hi-C).
Genomic ranges define both the x- and y-axes. Each pair of
bins is drawn as a rectangle, filled according to a value
matrix mapped through a color function.
}
\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqMatrix}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gr_x}}{A `GRanges` object for x-axis bins.}

\item{\code{gr_y}}{A `GRanges` object for y-axis bins.}

\item{\code{value}}{A numeric matrix of size length(gr_y) × length(gr_x).}

\item{\code{col_fun}}{A function mapping numeric values to colors
(e.g. from circlize::colorRamp2).}

\item{\code{yCol}}{y column}

\item{\code{y_vals}}{y column}

\item{\code{coordOriginal}}{List storing original GRanges and values.}

\item{\code{coordCanvas}}{List of per-panel cell rectangles (x0, x1, y0, y1, color).}

\item{\code{aesthetics}}{Plot aesthetics (border, lwd).}

\item{\code{defaultAesthetics}}{Default aesthetics for tiles.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqMatrix-new}{\code{SeqMatrix$new()}}
\item \href{#method-SeqMatrix-prep}{\code{SeqMatrix$prep()}}
\item \href{#method-SeqMatrix-draw}{\code{SeqMatrix$draw()}}
\item \href{#method-SeqMatrix-clone}{\code{SeqMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqMatrix-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SeqMatrix` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqMatrix$new(
  gr_x,
  gr_y = NULL,
  value,
  col_fun,
  yCol = NULL,
  y_vals = NULL,
  aesthetics = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr_x}}{GRanges for x-axis bins.}

\item{\code{gr_y}}{GRanges for y-axis bins (if NULL, use gr_x).}

\item{\code{value}}{Numeric matrix (nrow = length(gr_y), ncol = length(gr_x)).}

\item{\code{col_fun}}{Function mapping numeric values to colors.}

\item{\code{yCol}}{y column}

\item{\code{y_vals}}{y values}

\item{\code{aesthetics}}{List of aesthetics (border, lwd).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqMatrix-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqMatrix-prep}{}}}
\subsection{Method \code{prep()}}{
Prepare heatmap cell coordinates for plotting in canvas space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqMatrix$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{List of panel metadata for the current track.}

\item{\code{track_windows}}{A `GRanges` of genomic windows for the track.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqMatrix-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqMatrix-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the heatmap tiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqMatrix$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
