% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqPlot}
\alias{SeqPlot}
\title{SeqPlot Class}
\format{
A \code{SeqPlot} object with the following fields:
}
\value{
A new \code{SeqPlot} object.
}
\description{
TODO: Describe the SeqPlot class.
}
\section{Methods}{

\describe{
  \item{\code{new(tracks)}}{Create a new \code{SeqPlot} object.}
  \item{\code{layoutGrid()}}{TODO: Describe 'layoutGrid' method.}
  \item{\code{drawGrid()}}{TODO: Describe 'drawGrid' method.}
  \item{\code{drawAxes()}}{TODO: Describe 'drawAxes' method.}
  \item{\code{drawElements()}}{TODO: Describe 'drawElements' method.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{arc_track_idx}}{TODO: Describe the 'arc_track_idx' field.}

\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{defaultAesthetics}}{TODO: Describe the 'defaultAesthetics' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{full}}{TODO: Describe the 'full' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{hjust}}{TODO: Describe the 'hjust' field.}

\item{\code{inner}}{TODO: Describe the 'inner' field.}

\item{\code{just}}{TODO: Describe the 'just' field.}

\item{\code{label}}{TODO: Describe the 'label' field.}

\item{\code{layout}}{TODO: Describe the 'layout' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{panelBounds}}{TODO: Describe the 'panelBounds' field.}

\item{\code{rot}}{TODO: Describe the 'rot' field.}

\item{\code{track}}{TODO: Describe the 'track' field.}

\item{\code{trackAesthetics}}{TODO: Describe the 'trackAesthetics' field.}

\item{\code{trackBackground}}{TODO: Describe the 'trackBackground' field.}

\item{\code{trackBounds}}{TODO: Describe the 'trackBounds' field.}

\item{\code{trackHeights}}{TODO: Describe the 'trackHeights' field.}

\item{\code{track_windows_list}}{TODO: Describe the 'track_windows_list' field.}

\item{\code{tracks}}{TODO: Describe the 'tracks' field.}

\item{\code{vjust}}{TODO: Describe the 'vjust' field.}

\item{\code{width}}{TODO: Describe the 'width' field.}

\item{\code{window}}{TODO: Describe the 'window' field.}

\item{\code{windows}}{TODO: Describe the 'windows' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{x1}}{TODO: Describe the 'x1' field.}

\item{\code{xAxisLabels}}{TODO: Describe the 'xAxisLabels' field.}

\item{\code{xAxisLine}}{TODO: Describe the 'xAxisLine' field.}

\item{\code{xScaleFactor}}{TODO: Describe the 'xScaleFactor' field.}

\item{\code{xscale}}{TODO: Describe the 'xscale' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}

\item{\code{yscale}}{TODO: Describe the 'yscale' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{arc_track_idx}}{TODO: Describe the 'arc_track_idx' field.}

\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{full}}{TODO: Describe the 'full' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{hjust}}{TODO: Describe the 'hjust' field.}

\item{\code{inner}}{TODO: Describe the 'inner' field.}

\item{\code{just}}{TODO: Describe the 'just' field.}

\item{\code{label}}{TODO: Describe the 'label' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{panelBounds}}{TODO: Describe the 'panelBounds' field.}

\item{\code{rot}}{TODO: Describe the 'rot' field.}

\item{\code{track}}{TODO: Describe the 'track' field.}

\item{\code{trackAesthetics}}{TODO: Describe the 'trackAesthetics' field.}

\item{\code{trackBackground}}{TODO: Describe the 'trackBackground' field.}

\item{\code{trackBounds}}{TODO: Describe the 'trackBounds' field.}

\item{\code{trackHeights}}{TODO: Describe the 'trackHeights' field.}

\item{\code{track_windows_list}}{TODO: Describe the 'track_windows_list' field.}

\item{\code{vjust}}{TODO: Describe the 'vjust' field.}

\item{\code{width}}{TODO: Describe the 'width' field.}

\item{\code{window}}{TODO: Describe the 'window' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{x1}}{TODO: Describe the 'x1' field.}

\item{\code{xAxisLabels}}{TODO: Describe the 'xAxisLabels' field.}

\item{\code{xAxisLine}}{TODO: Describe the 'xAxisLine' field.}

\item{\code{xScaleFactor}}{TODO: Describe the 'xScaleFactor' field.}

\item{\code{xscale}}{TODO: Describe the 'xscale' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}

\item{\code{yscale}}{TODO: Describe the 'yscale' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqPlot-new}{\code{SeqPlot$new()}}
\item \href{#method-SeqPlot-layoutGrid}{\code{SeqPlot$layoutGrid()}}
\item \href{#method-SeqPlot-drawGrid}{\code{SeqPlot$drawGrid()}}
\item \href{#method-SeqPlot-drawAxes}{\code{SeqPlot$drawAxes()}}
\item \href{#method-SeqPlot-drawElements}{\code{SeqPlot$drawElements()}}
\item \href{#method-SeqPlot-clone}{\code{SeqPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPlot$new(
  tracks = list(),
  windows = defaultGenomeWindows(),
  layout = NULL,
  aesthetics = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tracks}}{TODO: Describe parameter 'tracks'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPlot-layoutGrid"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPlot-layoutGrid}{}}}
\subsection{Method \code{layoutGrid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPlot$layoutGrid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPlot-drawGrid"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPlot-drawGrid}{}}}
\subsection{Method \code{drawGrid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPlot$drawGrid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPlot-drawAxes"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPlot-drawAxes}{}}}
\subsection{Method \code{drawAxes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPlot$drawAxes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPlot-drawElements"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPlot-drawElements}{}}}
\subsection{Method \code{drawElements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPlot$drawElements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
