% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqPoint}
\alias{SeqPoint}
\title{SeqPoint}
\usage{
SeqPoint(gr, yCol = NULL, colorCol = NULL, color = "black")
}
\arguments{
\item{gr}{GRanges object}

\item{yCol}{Metadata column for y-axis value}

\item{colorCol}{Metadata column for color (optional)}

\item{color}{Default color}
}
\description{
Constructs a SeqPoint object for plotting point features.
}
