% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqPoint}
\alias{SeqPoint}
\title{SeqPoint Class}
\format{
A \code{SeqPoint} object with the following fields:
}
\value{
A new \code{SeqPoint} object.
}
\description{
TODO: Describe the SeqPoint class.
}
\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqPoint}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{coordOriginal}}{TODO: Describe the 'coordOriginal' field.}

\item{\code{defaultAesthetics}}{TODO: Describe the 'defaultAesthetics' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqPoint-new}{\code{SeqPoint$new()}}
\item \href{#method-SeqPoint-prep}{\code{SeqPoint$prep()}}
\item \href{#method-SeqPoint-draw}{\code{SeqPoint$draw()}}
\item \href{#method-SeqPoint-clone}{\code{SeqPoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoint-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPoint$new(gr, yCol = NULL, aesthetics = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoint-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoint-prep}{}}}
\subsection{Method \code{prep()}}{
Preps input data for this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPoint$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoint-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoint-draw}{}}}
\subsection{Method \code{draw()}}{
Preps input data for this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPoint$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqPoint-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqPoint-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqPoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
