% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqRecon}
\alias{SeqRecon}
\title{SeqRecon Class}
\format{
A \code{SeqRecon} object with the following fields:
}
\value{
A new \code{SeqRecon} object.
}
\description{
TODO: Describe the SeqRecon class.
}
\section{Methods}{

\describe{
  \item{\code{prep(layout_all_tracks, track_windows_list, arc_track_idx)}}{TODO: Describe 'prep' method.}
  \item{\code{draw()}}{TODO: Describe 'draw' method.}
}
}

\section{Super classes}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{\link[THEfunc:SeqLink]{THEfunc::SeqLink}} -> \code{\link[THEfunc:SeqArch]{THEfunc::SeqArch}} -> \code{SeqRecon}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{arc_track_idx}}{TODO: Describe the 'arc_track_idx' field.}

\item{\code{col_del}}{TODO: Describe the 'col_del' field.}

\item{\code{col_dup}}{TODO: Describe the 'col_dup' field.}

\item{\code{col_h2h}}{TODO: Describe the 'col_h2h' field.}

\item{\code{col_t2t}}{TODO: Describe the 'col_t2t' field.}

\item{\code{col_trans}}{TODO: Describe the 'col_trans' field.}

\item{\code{curve}}{TODO: Describe the 'curve' field.}

\item{\code{gr1}}{TODO: Describe the 'gr1' field.}

\item{\code{gr2}}{TODO: Describe the 'gr2' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}

\item{\code{last_arc_track}}{TODO: Describe the 'last_arc_track' field.}

\item{\code{layout_all_tracks}}{TODO: Describe the 'layout_all_tracks' field.}

\item{\code{orientation}}{TODO: Describe the 'orientation' field.}

\item{\code{t0}}{TODO: Describe the 't0' field.}

\item{\code{t1}}{TODO: Describe the 't1' field.}

\item{\code{tierMultipliers}}{TODO: Describe the 'tierMultipliers' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{curve}}{TODO: Describe the 'curve' field.}

\item{\code{gr1}}{TODO: Describe the 'gr1' field.}

\item{\code{gr2}}{TODO: Describe the 'gr2' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{initialize}}{TODO: Describe the 'initialize' field.}

\item{\code{orientation}}{TODO: Describe the 'orientation' field.}

\item{\code{t0}}{TODO: Describe the 't0' field.}

\item{\code{t1}}{TODO: Describe the 't1' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqRecon-new}{\code{SeqRecon$new()}}
\item \href{#method-SeqRecon-prep}{\code{SeqRecon$prep()}}
\item \href{#method-SeqRecon-draw}{\code{SeqRecon$draw()}}
\item \href{#method-SeqRecon-clone}{\code{SeqRecon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRecon-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRecon-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRecon$new(
  gr1,
  gr2,
  yCol = NULL,
  t0 = 0,
  t1 = 0,
  orientation = "*",
  curve = "length",
  aesthetics = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRecon-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRecon-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRecon$prep(layout_all_tracks, track_windows_list, arc_track_idx)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRecon-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRecon-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRecon$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRecon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRecon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRecon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
