% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqRecon}
\alias{SeqRecon}
\title{SeqRecon}
\usage{
SeqRecon(
  gr1,
  gr2,
  orientation = "*",
  orientationCol = NULL,
  color = NULL,
  colorCol = NULL,
  y0 = 0,
  y1 = 0,
  t0 = 0,
  t1 = 0,
  curve = "equal"
)
}
\description{
Constructs a ReCon-style arc from paired breakpoints.
}
