% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqRect}
\alias{SeqRect}
\title{SeqRect Class}
\format{
A \code{SeqRect} object with the following fields:
}
\value{
A new \code{SeqRect} object.
}
\description{
TODO: Describe the SeqRect class.
}
\section{Methods}{

\describe{
  \item{\code{new(gr, yCol, aesthetics)}}{Create a new \code{SeqRect} object.}
  \item{\code{prep(layout_track, track_windows)}}{TODO: Describe 'prep' method.}
  \item{\code{draw()}}{TODO: Describe 'draw' method.}
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqRect}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{defaultAesthetics}}{TODO: Describe the 'defaultAesthetics' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{width}}{TODO: Describe the 'width' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{x1}}{TODO: Describe the 'x1' field.}

\item{\code{y}}{TODO: Describe the 'y' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{fill}}{TODO: Describe the 'fill' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{height}}{TODO: Describe the 'height' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{width}}{TODO: Describe the 'width' field.}

\item{\code{x}}{TODO: Describe the 'x' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{x1}}{TODO: Describe the 'x1' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqRect-new}{\code{SeqRect$new()}}
\item \href{#method-SeqRect-prep}{\code{SeqRect$prep()}}
\item \href{#method-SeqRect-draw}{\code{SeqRect$draw()}}
\item \href{#method-SeqRect-clone}{\code{SeqRect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$new(gr, yCol = NULL, aesthetics = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{TODO: Describe parameter 'gr'.}

\item{\code{yCol}}{TODO: Describe parameter 'yCol'.}

\item{\code{aesthetics}}{TODO: Describe parameter 'aesthetics'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
