% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqRect}
\alias{SeqRect}
\title{SeqRect R6 Class}
\description{
R6 class for drawing rectangular genomic features (boxes) in the SeqPlot
framework. Inherits from [SeqElement].
}
\details{
`SeqRect` is useful for visualizing genomic intervals as filled rectangles,
for example in bar plots, ideograms, or feature blocks. By default, each
rectangle is drawn centered on a y-value (`yCol` or a constant) with a
configurable relative height (`width` aesthetic).
}
\examples{
gr <- GenomicRanges::GRanges(
  "chr1",
  IRanges::IRanges(c(1, 100), width = 50),
  score = c(0.2, 0.8)
)
rects <- SeqRect$new(gr, yCol = "score")
rects$prep(layout_track = some_layout, track_windows = some_windows)
rects$draw()

}
\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqRect}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing genomic intervals.}

\item{\code{y0}}{Numeric vector of lower rectangle y-values (computed in prep).}

\item{\code{y1}}{Numeric vector of upper rectangle y-values (computed in prep).}

\item{\code{y}}{Numeric vector of rectangle center y-values.}

\item{\code{yCol}}{Optional column name in `gr` used for y-values.}

\item{\code{coordCanvas}}{A list of matrices containing transformed rectangle
coordinates for each window (x0, x1, y0, y1).}

\item{\code{aesthetics}}{List of current aesthetics merged with defaults.}

\item{\code{defaultAesthetics}}{Default aesthetics: \code{fill = "grey80"},
\code{col = "#1C1B1A"}, \code{lwd = 0.5}, \code{width = 0.1}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqRect-new}{\code{SeqRect$new()}}
\item \href{#method-SeqRect-prep}{\code{SeqRect$prep()}}
\item \href{#method-SeqRect-draw}{\code{SeqRect$draw()}}
\item \href{#method-SeqRect-clone}{\code{SeqRect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SeqRect` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$new(gr, yCol = NULL, aesthetics = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing genomic intervals.}

\item{\code{yCol}}{Optional column name in `gr` used for rectangle y-centers.}

\item{\code{aesthetics}}{Optional list of aesthetic overrides.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SeqRect` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-prep}{}}}
\subsection{Method \code{prep()}}{
Prepare rectangle coordinates by mapping genomic intervals into
panel-relative canvas space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{A list of panel layout metadata.}

\item{\code{track_windows}}{A `GRanges` object of genomic windows.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-draw}{}}}
\subsection{Method \code{draw()}}{
Draw rectangles on the plotting canvas.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqRect-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqRect-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqRect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
