% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\docType{class}
\name{SeqSegment}
\alias{SeqSegment}
\title{SeqSegment Class}
\format{
A \code{SeqSegment} object with the following fields:
}
\value{
A new \code{SeqSegment} object.
}
\description{
TODO: Describe the SeqSegment class.
}
\section{Methods}{

\describe{
  \item{\code{new(gr, yCol, y0Col, y1Col, aesthetics)}}{Create a new \code{SeqSegment} object.}
  \item{\code{prep(layout_track, track_windows)}}{TODO: Describe 'prep' method.}
  \item{\code{draw()}}{TODO: Describe 'draw' method.}
}
}

\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqSegment}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aesthetics}}{TODO: Describe the 'aesthetics' field.}

\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{coordCanvas}}{TODO: Describe the 'coordCanvas' field.}

\item{\code{coordOriginal}}{TODO: Describe the 'coordOriginal' field.}

\item{\code{defaultAesthetics}}{TODO: Describe the 'defaultAesthetics' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{gr}}{TODO: Describe the 'gr' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{x1}}{TODO: Describe the 'x1' field.}

\item{\code{y0}}{TODO: Describe the 'y0' field.}

\item{\code{y0Col}}{TODO: Describe the 'y0Col' field.}

\item{\code{y1}}{TODO: Describe the 'y1' field.}

\item{\code{y1Col}}{TODO: Describe the 'y1Col' field.}

\item{\code{yCol}}{TODO: Describe the 'yCol' field.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{col}}{TODO: Describe the 'col' field.}

\item{\code{gp}}{TODO: Describe the 'gp' field.}

\item{\code{lwd}}{TODO: Describe the 'lwd' field.}

\item{\code{x0}}{TODO: Describe the 'x0' field.}

\item{\code{x1}}{TODO: Describe the 'x1' field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqSegment-new}{\code{SeqSegment$new()}}
\item \href{#method-SeqSegment-prep}{\code{SeqSegment$prep()}}
\item \href{#method-SeqSegment-draw}{\code{SeqSegment$draw()}}
\item \href{#method-SeqSegment-clone}{\code{SeqSegment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegment$new(
  gr,
  yCol = NULL,
  y0Col = NULL,
  y1Col = NULL,
  aesthetics = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{TODO: Describe parameter 'gr'.}

\item{\code{yCol}}{TODO: Describe parameter 'yCol'.}

\item{\code{y0Col}}{TODO: Describe parameter 'y0Col'.}

\item{\code{y1Col}}{TODO: Describe parameter 'y1Col'.}

\item{\code{aesthetics}}{TODO: Describe parameter 'aesthetics'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-prep}{}}}
\subsection{Method \code{prep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegment$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegment$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqSegment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqSegment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqSegment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
