% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqSegment}
\alias{SeqSegment}
\title{SeqSegment}
\usage{
SeqSegment(gr, yCol = NULL, y1Col = NULL, colorCol = NULL, color = "black")
}
\arguments{
\item{gr}{GRanges object}

\item{yCol}{Metadata column for start y-value}

\item{y1Col}{Metadata column for end y-value}

\item{colorCol}{Metadata column for color (optional)}

\item{color}{Default color}
}
\description{
Constructs a SeqSegment object for plotting genomic segments.
}
