% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqString}
\alias{SeqString}
\title{SeqString}
\usage{
SeqString(
  df,
  t0 = 0,
  t1 = 0,
  color = "black",
  colorCol = NULL,
  orientation = "*",
  orientationCol = "orientation"
)
}
\arguments{
\item{df}{A data.frame or DataFrame with columns chr0, x0, y0, chr1, x1, y1, orientation}
}
\description{
Constructs a SeqString object from a BED-like input
}
