% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqTile}
\alias{SeqTile}
\title{SeqTile R6 Class}
\description{
An R6 class for plotting categorical heatmap-style tiles.
Genomic ranges define the x-axis, and a categorical grouping column
defines the y-axis. Each tile is filled with the color stored in the
`color` metadata column of the GRanges.
}
\section{Super class}{
\code{\link[THEfunc:SeqElement]{THEfunc::SeqElement}} -> \code{SeqTile}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gr}}{A `GRanges` object containing genomic ranges
and a required `color` metadata column.}

\item{\code{groupCol}}{Name of metadata column in `gr` used
to define grouping (y-axis categories).}

\item{\code{groups}}{Vector of unique group labels, in the order
they appear in the input data.}

\item{\code{coordOriginal}}{Original GRanges object.}

\item{\code{coordCanvas}}{List of per-panel tile coordinates (x0, x1, y0, y1, color).}

\item{\code{aesthetics}}{Plotting aesthetics (border, lwd).}

\item{\code{defaultAesthetics}}{Default non-fill aesthetics for tiles.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqTile-new}{\code{SeqTile$new()}}
\item \href{#method-SeqTile-prep}{\code{SeqTile$prep()}}
\item \href{#method-SeqTile-draw}{\code{SeqTile$draw()}}
\item \href{#method-SeqTile-clone}{\code{SeqTile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqTile-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqTile-new}{}}}
\subsection{Method \code{new()}}{
Create a new `SeqTile` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTile$new(gr, groupCol, aesthetics = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gr}}{A `GRanges` object with a `color` metadata column.}

\item{\code{groupCol}}{Column name in `gr` metadata for grouping (y-axis).}

\item{\code{aesthetics}}{List of aesthetics (`border`, `lwd`).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqTile-prep"></a>}}
\if{latex}{\out{\hypertarget{method-SeqTile-prep}{}}}
\subsection{Method \code{prep()}}{
Prepare tile coordinates for plotting in canvas space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTile$prep(layout_track, track_windows)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout_track}}{A list of panel metadata for the current track.}

\item{\code{track_windows}}{A `GRanges` of genomic windows for the track.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqTile-draw"></a>}}
\if{latex}{\out{\hypertarget{method-SeqTile-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the tiles using grid graphics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTile$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqTile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqTile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqTile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
