% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqPlot.R
\name{SeqTrack}
\alias{SeqTrack}
\title{SeqTrack}
\usage{
SeqTrack(
  featureList,
  secondaryAxis = FALSE,
  xAxis = FALSE,
  yAxis = TRUE,
  yTitle = "",
  trackBackgroundColor = "transparent",
  trackBorderColor = "transparent",
  windowBackgroundColor = "transparent",
  windowBorderColor = "grey50",
  fixedStubs = T
)
}
\arguments{
\item{featureList}{List of feature objects.}

\item{secondaryAxis}{Logical; draw a second y-axis.}

\item{xAxis}{Logical; draw an x-axis.}

\item{yAxis}{Logical; draw a y-axis.}

\item{yTitle}{Character; title of the y-axis.}
}
\value{
A SeqTrack object.
}
\description{
Combines multiple SeqFeature, SeqLink, or SeqAnnotation objects into a track.
}
